# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VisiteCRM API - SaaS CRM for tourism agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List marketplace products
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: active
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── ORDERS ───────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── CAMPAIGNS ────────────────────────────────────────────────────
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List marketing campaigns
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── NPS ──────────────────────────────────────────────────────────
  /nps:
    get:
      operationId: listNpsResponses
      tags: [nps]
      summary: List NPS responses
      parameters:
        - name: classification
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: NPS responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NpsResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ─── DASHBOARD ─────────────────────────────────────────────────
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
        price:
          type: number
        promotionalPrice:
          type: ["number", "null"]
        stock:
          type: ["integer", "null"]
        active:
          type: boolean
        featured:
          type: boolean
        createdAt:
          type: string
      required: [id, name, slug, type, price, active, featured, createdAt]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
        price:
          type: number
        promotionalPrice:
          type: ["number", "null"]
        stock:
          type: ["integer", "null"]
      required: [name, type, price]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        promotionalPrice:
          type: ["number", "null"]
        active:
          type: ["boolean", "null"]
        featured:
          type: ["boolean", "null"]
        stock:
          type: ["integer", "null"]

    # ─── ORDER ─────────────────────────────────────────────────────
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        totalAmount:
          type: number
        finalAmount:
          type: number
        status:
          type: string
        paymentStatus:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [id, userId, totalAmount, finalAmount, status, paymentStatus, createdAt, items]
    OrderItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        quantity:
          type: integer
        price:
          type: number
        productName:
          type: ["string", "null"]
      required: [id, productId, quantity, price]
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
        paymentStatus:
          type: ["string", "null"]

    # ─── CAMPAIGN ──────────────────────────────────────────────────
    Campaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        status:
          type: string
        subject:
          type: ["string", "null"]
        content:
          type: string
        scheduledAt:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        recipientsCount:
          type: integer
        sentCount:
          type: integer
        openedCount:
          type: integer
        clickedCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, type, status, content, recipientsCount, sentCount, openedCount, clickedCount, createdAt]
    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        subject:
          type: ["string", "null"]
        content:
          type: string
        targetSegment:
          type: object
          additionalProperties: true
        scheduledAt:
          type: ["string", "null"]
      required: [name, type, content, targetSegment]
    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]

    # ─── NPS ───────────────────────────────────────────────────────
    NpsResponse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        score:
          type: integer
        classification:
          type: string
        feedback:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, score, classification, createdAt]
...